+++ crates/networking/p2p/peer_handler.rs
@@ -545,13 +545,13 @@ impl PeerHandler {
     /// - The requested peer did not return a valid response in the given time limit
     async fn request_block_bodies_inner(
         &mut self,
-        block_hashes: Vec<H256>,
+        block_hashes: &[H256],
     ) -> Result<Option<(Vec<BlockBody>, H256)>, PeerHandlerError> {
         let block_hashes_len = block_hashes.len();
         let request_id = rand::random();
         let request = RLPxMessage::GetBlockBodies(GetBlockBodies {
             id: request_id,
-            block_hashes: block_hashes.clone(),
+            block_hashes: block_hashes.to_vec(),
         });
         match self.get_random_peer(&SUPPORTED_ETH_CAPABILITIES).await? {
             None => Ok(None),
@@ -609,13 +609,10 @@ impl PeerHandler {
     /// - No peer returned a valid response in the given time and retry limits
     pub async fn request_block_bodies(
         &mut self,
-        block_hashes: Vec<H256>,
+        block_hashes: &[H256],
     ) -> Result<Option<Vec<BlockBody>>, PeerHandlerError> {
         for _ in 0..REQUEST_RETRY_ATTEMPTS {
-            if let Some((block_bodies, _)) = self
-                .request_block_bodies_inner(block_hashes.clone())
-                .await?
-            {
+            if let Some((block_bodies, _)) = self.request_block_bodies_inner(block_hashes).await? {
                 return Ok(Some(block_bodies));
             }
         }
@@ -634,9 +631,8 @@ impl PeerHandler {
         let block_hashes: Vec<H256> = block_headers.iter().map(|h| h.hash()).collect();
 
         for _ in 0..REQUEST_RETRY_ATTEMPTS {
-            let Some((block_bodies, peer_id)) = self
-                .request_block_bodies_inner(block_hashes.clone())
-                .await?
+            let Some((block_bodies, peer_id)) =
+                self.request_block_bodies_inner(&block_hashes).await?
             else {
                 continue; // Retry on empty response
             };